requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
streamlit>=1.40.0
langchain>=0.3.0
langchain-groq>=0.1.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=2.0.0
requests>=2.31.0